refactor: move stores back to stores folder
