+++ src/main.ts
@@ -2,7 +2,7 @@ import './assets/main.css'
 import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import store from './store'
+import store from './stores'
 
 const app = createApp(App)
 app.use(store)
