+++ src/stores/warehouseStore.ts
